src/requirements.txt
fastapi
uvicorn
apscheduler
jinja2
httpx
sqlite-utils
python-multipart
python-dotenv
authlib
python-dotenv
python-multipart
itsdangerous